requirements_amd.txt
# AMD ROCm GPU 依赖文件
# ===================================================================================
# 说明:
# - AMD ROCm PyTorch 不在本 requirements 文件中，由 packaging/launch.py 单独安装
# - Windows 上使用 ROCm 7.2 固定 URL 两阶段安装:
#   1) 先安装 ROCm SDK 依赖:
#      rocm_sdk_core / rocm_sdk_devel / rocm_sdk_libraries_custom / rocm
#   2) 再安装 PyTorch 三件套:
#      torch / torchaudio / torchvision
#   下载地址（ROCm SDK）:
#      https://repo.radeon.com/rocm/windows/rocm-rel-7.2/rocm_sdk_core-7.2.0.dev0-py3-none-win_amd64.whl
#      https://repo.radeon.com/rocm/windows/rocm-rel-7.2/rocm_sdk_devel-7.2.0.dev0-py3-none-win_amd64.whl
#      https://repo.radeon.com/rocm/windows/rocm-rel-7.2/rocm_sdk_libraries_custom-7.2.0.dev0-py3-none-win_amd64.whl
#      https://repo.radeon.com/rocm/windows/rocm-rel-7.2/rocm-7.2.0.dev0.tar.gz
#   下载地址（PyTorch）:
#      https://repo.radeon.com/rocm/windows/rocm-rel-7.2/torch-2.9.1%2Brocmsdk20260116-cp312-cp312-win_amd64.whl
#      https://repo.radeon.com/rocm/windows/rocm-rel-7.2/torchaudio-2.9.1%2Brocmsdk20260116-cp312-cp312-win_amd64.whl
#      https://repo.radeon.com/rocm/windows/rocm-rel-7.2/torchvision-0.24.1%2Brocmsdk20260116-cp312-cp312-win_amd64.whl
#
# 前置要求:
# - Windows 版 ROCm 7.2 PyTorch 必须安装 AMD 显卡驱动 26.1.1
# - Python 版本必须为 3.12
#
# 当前支持的 AMD 显卡（仅以下型号）:
# - MI300A / MI300X                     -> gfx94X-dcgpu
# - MI350X / MI355X                     -> gfx950-dcgpu
# - AMD RX 7900 XTX                     -> gfx110X-dgpu
# - AMD RX 7800 XT                      -> gfx110X-dgpu
# - AMD RX 7700S / Framework Laptop 16  -> gfx110X-dgpu
# - AMD Strix Halo iGPU                 -> gfx1151
# - AMD RX 9060 / RX 9060 XT            -> gfx120X-all
# - AMD RX 9070 / RX 9070 XT            -> gfx120X-all
#
# 常见不支持:
# - RX 5000 系列 (gfx101X-dgpu)
# - RX 6000 系列 (gfx103X-dgpu)
# - Vega / Radeon VII (gfx90X-dcgpu)
#
# 注意事项:
# - AMD ROCm PyTorch 是实验性功能，稳定性可能不如 NVIDIA CUDA 版本
# - 首次运行可能需要编译某些操作，请耐心等待
# - 如果遇到问题，建议使用 CPU 版本 (requirements_cpu.txt)
# - AMD ROCm 对 Windows 的支持有限，Linux 下体验更好
# ===================================================================================


aiofiles==25.1.0
aiohttp==3.12.15
aioshutil==1.5
albumentations==2.0.8
arabic-reshaper==3.0.0
backports.statistics==0.1.0
colorama==0.4.6
ctranslate2>=4.6.0
customtkinter==5.2.2
einops==0.8.1
fastapi>=0.116.1
fire==0.7.1
freetype-py==2.5.1
fugashi==1.3.2
python-multipart==0.0.20

httpcore==1.0.9
httpx>=0.28.1,<1.0.0
huggingface-hub>=0.36.0
ijson==3.4.0
jaconv==0.3.4
kornia==0.8.1
langcodes==3.5.0
loguru==0.7.3
matplotlib==3.10.5
networkx==3.5
# numpy 由 PyTorch 自动管理，不要在此显式指定以避免版本冲突
omegaconf==2.3.0
onnxruntime>=1.23.0
openai==2.14.0
curl_cffi>=0.7.0
opencv-contrib-python>=4.10.0
packaging<25.0
pandas==2.3.1
Pillow==11.3.0
pillow-heif==0.21.0
psutil==5.9.5
py3langid==0.2.2
pyclipper==1.3.0.post6
pydantic==2.11.7
pyperclip==1.9.0
PyHyphen==4.0.4
PyQt6==6.9.1
pytest==7.2.2
python-dotenv==1.1.1
rich>=14.1.0
safetensors==0.6.2
spandrel==0.4.1
scipy==1.16.1
# YOLO OBB checkpoint 反序列化需要与模型导出版本对齐的 ultralytics
ultralytics==8.4.14
# sentencepiece 已移除，改用 tokenizers 库 (use_fast=True)
setuptools==80.9.0
Shapely==2.1.1
scikit-image==0.25.2
starlette==0.47.2
tiktoken==0.11.0
timm>=1.0.19
transformers==4.55.0
accelerate>=1.12.0
tqdm==4.67.1
unidic_lite==1.0.8
uvicorn==0.35.0
websockets==14.2
anyio>=4.8.0
# xformers - AMD ROCm 支持有限,暂不包含
google-genai==1.57.0
# pydensecrf - 按平台选择安装来源
# Windows: 使用预编译 wheel（避免本地编译失败）
pydensecrf @ https://github.com/hgmzhn/manga-translator-ui/releases/download/v1.7.6/pydensecrf-1.0-cp312-cp312-win_amd64.whl ; sys_platform == 'win32'
# Linux/macOS: 从源码安装（需要 C++ 编译工具）
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git ; sys_platform != 'win32'
# 如果 Windows 下载失败，可尝试镜像站：
# pydensecrf @ https://gh-proxy.com/https://github.com/hgmzhn/manga-translator-ui/releases/download/v1.7.6/pydensecrf-1.0-cp312-cp312-win_amd64.whl ; sys_platform == 'win32'
torchsummary
open_clip_torch==3.1.0
python-bidi==0.6.6
tufup==0.9.0
google-cloud==0.34.0


# PDF/EPUB/CBZ 压缩包支持
PyMuPDF==1.25.5
rarfile==4.2
py7zr==0.22.0

# Server 模块依赖
bcrypt==4.2.1
cryptography==44.0.0
APScheduler==3.10.4
pytest-anyio==0.0.0
